Documentation: Briefly document the mail-based contribution process
While familiarity with git is nowadays common, many developers are not
familiar with mail-based patch submission. Add a paragraph to the
contribution guide to mention that libcamera uses a mail-based process
(which seems to have been taken for granted so much that we forgot to
say it explicitly), and link to git-send-email.io to help newcomers
setup git-send-email.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo@jmondi.org>