test: v4l2_videodevice: Fix format configuration in the vimc pipeline
The V4L2VideoDeviceTest class configures the capture pipeline with
parameters that are partly hardcoded, and partly come from the current
configuration of the device. In particular, with the vimc pipeline, the
sensor subdevice is configured with the size retrieved from the capture
video node, and the video node is then reconfigured to 640x480.

Relying on the current (and thus possibly random) device configuration
can lead to broken pipes when starting streaming. This currently causes
failures of the dequeue_watchdog test when run after the formats test.

Fix it by explicitly setting the same size for both the vimc subdevs and
the video capture device.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Tested-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>