libcamera: Add a SensorSensitivity property
The SensorSensitivity property is a scaling factor that describes how sensitive
the selected sensor mode is compared to other readout modes of the same sensor.
For example, a binned mode might have twice the sensitivity of the full
resolution mode, meaning you would get double the signal level for the same
exposure and gains.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo@jmondi.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>