cam: Add a parser for capture scripts
Add a parser class to the cam test application to control the capture
operations through a yaml script.

The capture script currently allow to specify a list of controls and
their associated values to be applied per-frame.

Also add a trivial capture script example to showcase the intended
script structure.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo@jmondi.org>